# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.DetectEntryRequest:
    properties:
      conf:
        type: number
      device:
        type: string
      imageId:
        type: string
      iou:
        type: number
      modelName:
        type: string
      modelVersion:
        type: string
      weights:
        type: string
    type: object
...
  handlers.importRoomsRequest:
    properties:
      rooms:
        items:
          properties:
            description:
              type: string
            id:
              type: string
            model_path:
              type: string
            name:
              type: string
            points:
              items:
                properties:
                  code:
                    type: string
                  id:
                    type: string
                  location:
                    type: string
                  name:
                    type: string
                type: object
              type: array
            status:
              type: string
          type: object
        type: array
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@foreignscan.com
    name: API Support
    url: http://www.foreignscan.com/support
  description: ForeignScan后端API服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ForeignScan API
  version: "1.0"
paths:
  /detect:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.DetectEntryRequest'
      produces:
      - application/json
      responses:
        "202":
          description: 任务已启动，返回jobId与初始状态
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 兼容入口：前端直接传 imageId 触发单图推理
      tags:
      - detections
...
  /images:
    get:
      consumes:
      - application/json
      description: 获取系统中所有图片的列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取图片列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有图片列表
      tags:
      - images
  /images/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取特定图片的详细信息
      parameters:
      - description: 图片ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取图片详情
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 图片不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取单个图片详情
      tags:
      - images
...
  /images/by-date:
    get:
      consumes:
      - application/json
      description: 获取指定日期上传的所有图片
      parameters:
      - description: '日期 (格式: YYYY-MM-DD)'
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取图片列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 日期格式错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 根据日期获取图片
      tags:
      - images
  /images/filter:
    get:
      consumes:
      - application/json
      description: 支持按状态、房间ID、点位ID、时间范围（start/end）筛选图片。
      parameters:
      - description: 状态（已检测/未检测）
        in: query
        name: status
        type: string
      - description: 是否存在问题（true/false，仅在status=已检测时生效）
        in: query
        name: hasIssue
        type: string
      - description: 房间ID
        in: query
        name: roomId
        type: string
      - description: 点位ID
        in: query
        name: pointId
        type: string
      - description: 起始时间（YYYY-MM-DD 或 RFC3339）
        in: query
        name: start
        type: string
      - description: 结束时间（YYYY-MM-DD 或 RFC3339）
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取筛选结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 根据多条件筛选图片（状态/时间/房间/点位）
      tags:
      - images
  /predict:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 房间ID
        in: formData
        name: roomId
        required: true
        type: string
      - description: 点位ID
        in: formData
        name: pointId
        required: true
        type: string
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      - description: 置信度
        in: formData
        name: conf
        type: number
      - description: IoU
        in: formData
        name: iou
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 同步推理（房间+点位）
      tags:
      - detections
...
  /rooms/import:
    post:
      consumes:
      - application/json
      parameters:
      - description: 房间配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.importRoomsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 导入房间点位配置
      tags:
      - rooms
...
schemes:
- http
swagger: "2.0"